+++ app.js
@@ -1,0 +1,51 @@
+var express = require('express');
+var path = require('path');
+var logger = require('morgan');
+var compression = require('compression');
+var methodOverride = require('method-override');
+var session = require('express-session');
+var flash = require('express-flash');
+var bodyParser = require('body-parser');
+var expressValidator = require('express-validator');
+var dotenv = require('dotenv');
+var nunjucks = require('nunjucks');
+
+// Load environment variables from .env file
+dotenv.load();
+
+// Controllers
+var phoneController = require('./controllers/phone');
+
+var app = express();
+
+// view engine setup
+nunjucks.configure('views', {
+  autoescape: true,
+  express: app
+});
+app.set('view engine', 'html');
+app.set('port', process.env.PORT || 17738);
+app.use(compression());
+app.use(logger('dev'));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(expressValidator());
+app.use(methodOverride('_method'));
+app.use(session({ secret: process.env.SESSION_SECRET, resave: true, saveUninitialized: true }));
+app.use(flash());
+app.use(express.static(path.join(__dirname, 'public')));
+
+app.post('/new_phone_call', phoneController.newCall);
+app.get('/new_phone_call', phoneController.newCallTestGet);
+app.post('/redir_call_for_zip', phoneController.redirectCall);
+app.get('/redir_call_for_zip', phoneController.redirectCallTest);
+
+// Production error handler
+if (app.get('env') === 'production') {
+  app.use(function(err, req, res, next) {
+    console.error(err.stack);
+    res.sendStatus(err.status || 500);
+  });
+}
+
+module.exports = app;
